Remove `await` qualifier from `tmpPath` example
The rest of the examples in these docs for `tmpPath` indicate that it does not return a Promise. Additionally, the function itself also currently does not return a Promise.

If there are some changes planned here to make this async, feel free to disregard this PR.

Otherwise, this is just some tidying up to remove the `await` and reduce any future confusion.